+++ src/lang/builtins_string.ml
@@ -394,10 +394,14 @@ let () =
       Lang.string (Utils.interpolate (fun k -> List.assoc k l) s))
 
 let () =
-  add_builtin "string.quote" ~cat:String ~descr:"Escape shell metacharacters."
+  add_builtin "string.quote" ~cat:String
+    ~descr:
+      "Return a quoted copy of the given string, suitable for use as one \
+       argument in a command line, escaping all meta-characters. Warning: \
+       under Windows, the output is only suitable for use with programs that \
+       follow the standard Windows quoting conventions."
     [("", Lang.string_t, None, None)] Lang.string_t (fun p ->
-      let s = Lang.to_string (List.assoc "" p) in
-      Lang.string (Utils.quote s))
+      Lang.string (Filename.quote (Lang.to_string (List.assoc "" p))))
 
 let () =
   add_builtin "string.hex_of_int" ~cat:String